Forgot to move some object to types/checkers
